Move prometheus update endpoints
