+++ src/murfey/server/api/instrument.py
@@ -13,8 +13,8 @@ from sqlmodel import select
 from werkzeug.utils import secure_filename
 
 from murfey.server import sanitise
-from murfey.server.api import MurfeySessionID
 from murfey.server.api.auth import (
+    MurfeySessionID,
     create_access_token,
     instrument_server_tokens,
     oauth2_scheme,
